refactor(signing): route internal transport imports through signing/client sub-barrel
Previously each of mcp.ts / a2a.ts / mcp-tasks.ts / protocols/index.ts
imported auto-wiring helpers via deep paths
(`../signing/agent-fetch`, `../signing/agent-context`, `../signing/capability-priming`).
Now they import through `../signing/client` — the same path external
consumers use.

This means any future reorganization of the deep files (e.g., the file
consolidation #583 might do alongside its breaking cuts) only has to
update the sub-barrel, not every transport call-site.

No behavior change — sub-barrel re-exports the same symbols.
